+++ react/components/StarPicker.tsx
@@ -40,7 +40,7 @@ const StarPicker: FunctionComponent<StarPickerProps> = ({
   return (
     <label className={`${styles.starpicker} c-action-primary`}>
       {label && <span className={labelClasses}>{label}</span>}
-      <span className="t-heading-4">
+      <span className="t-heading-4 pointer">
         {stars.map((value, index) => (
           <Star
             key={index}